Move everything to `SelectV2` namespace
