+++ osu.Game/Screens/SelectV2/Footer/FooterButtonV2.cs
@@ -20,7 +20,7 @@ using osu.Game.Overlays;
 using osuTK;
 using osuTK.Graphics;
 
-namespace osu.Game.Screens.Select.FooterV2
+namespace osu.Game.Screens.SelectV2.Footer
 {
     public partial class FooterButtonV2 : OsuClickableContainer, IKeyBindingHandler<GlobalAction>
     {
